+++ oreClient/src/main/resources/assets/entries/home.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 
-const root = require('../Home.vue').default;
+const root = require('../pages/Home.vue').default;
 const app = new Vue({
     el: '#home',
     render: createElement => createElement(root),
