+++ oreClient/src/main/resources/assets/entries/project.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 
-const root = require('../components/Project.vue').default;
+const root = require('../pages/Project.vue').default;
 const app = new Vue({
     el: '#project',
     render: createElement => createElement(root, {
