+++ oreClient/src/main/resources/assets/entries/user-profile.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 
-const root = require('../UserProfile.vue').default;
+const root = require('../pages/UserProfile.vue').default;
 const app = new Vue({
     el: '#user-profile',
     render: createElement => createElement(root),
