+++ oreClient/src/main/resources/assets/entries/version-list.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 
-const root = require('../VersionList.vue').default;
+const root = require('../pages/VersionList.vue').default;
 const app = new Vue({
     el: '#version-list',
     render: createElement => createElement(root),
