+++ oreClient/src/main/resources/assets/pages/UserProfile.vue
@@ -4,7 +4,7 @@
 </template>
 
 <script>
-    import ProjectList from "./components/ProjectList";
+    import ProjectList from "./../components/ProjectList";
 
     export default {
         components: {
